restore original bidirectional_solver and rename improved version to bidirectional_solver_var1.py
